Add alpha support for adding history to alda commands
This is the replacement for the old alda append command, which is in
active developement. As such, this is subject to change, and untested.

To begin using the new features, use the 'alda-history-append'
commands. If your history gets corruped with invalid alda code, you
will be unable to keep playing files (since we will be sending invalid
history to the server). In this case, run 'alda-history-clear'.

See alda-lang/alda-client-java#1 for a discussion on alda-append and
history.